Entrega Final

+++ javascript/cart.js
@@ -1,6 +1,8 @@
 let productosEnCarrito = (localStorage.getItem("productos-en-carrito"));
 productosEnCarrito = JSON.parse(productosEnCarrito);
 
+//Dom
+
 const contenedorCarritoVacio = document.querySelector("#carrito-vacio");
 const contenedorCarritoProductos = document.querySelector("#carrito-productos");
 const contenedorCarritoAcciones = document.querySelector("#carrito-acciones");
@@ -11,7 +13,7 @@ const contenedorTotal = document.querySelector("#total");
 const botonComprar = document.querySelector("#carrito-acciones-comprar");
 
 
-
+// Productos del carrito
 
 function cargarProductosCarrito() {
 
@@ -71,18 +73,37 @@ function cargarProductosCarrito() {
 
 cargarProductosCarrito();
 
+// Actualizar los botones eliminar
 
 function actualizarBotonesEliminar() {
+    
     botonesEliminar = document.querySelectorAll(".carrito-producto-eliminar");
 
     botonesEliminar.forEach(boton => {
         boton.addEventListener("click", eliminarDelCarrito)
+        
     });
 }
 
+// Eliminar del carrito
+
 function eliminarDelCarrito(e) {
+
+    Toastify({
+
+        text: "Producto eliminado de su carrito.",
+        
+        duration: 3000,
+
+        style: {
+            background: "linear-gradient(to right, #f25c54, #ca3c25)",
+          }
+        
+        }).showToast();
+    
     const idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
+    
     
     productosEnCarrito.splice(index, 1);
     cargarProductosCarrito();
@@ -90,15 +111,42 @@ function eliminarDelCarrito(e) {
     localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
 }
 
+// Vaciar carrito
+
 botonVaciar.addEventListener("click", vaciarCarrito);
 
 function vaciarCarrito() {
+    Swal.fire({
+        title: '¿Estas seguro de vaciar tu carrito?',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+        confirmButtonText: 'Vaciar',
+        confirmButtonColor: '#1a7431',
+        cancelButtonColor: '#c81d25'
+    }).then((response => {
+        if(response.isConfirmed) {
 
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+            Toastify({
+
+                text: "Su carrito esta vacío.",
+                
+                duration: 3000,
+
+                style: {
+                    background: "linear-gradient(to right, #f25c54, #ca3c25)",
+                  }
+                
+                }).showToast();
+            
+        productosEnCarrito.length = 0;
+        localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+        cargarProductosCarrito();
+        } 
+    
+    }))
 
 }
+
 
 function actualizarTotal() {
     
@@ -107,18 +155,46 @@ function actualizarTotal() {
     total.innerText = `$${totalCalculado}`;
 }
 
+// Boton comprar
+
 botonComprar.addEventListener("click", comprarCarrito);
 
 function comprarCarrito() {
+    
+    Swal.fire({
+        title: 'Confirmar compra',
+        text: '¿Esta seguro de seguir con la compra?',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+        confirmButtonText: 'Comprar',
+        confirmButtonColor: '#1a7431',
+        cancelButtonColor: '#c81d25'
+    }).then((result) => {
 
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+        if(result.isConfirmed) {
+
+            Swal.fire({
+                title: 'Gracias por su compra',
+                showCancelButton: false,
+                showConfirmButtom: false,
+                confirmButtonText: 'Cerrar',
+                confirmButtonColor: '#2364aa',
+            })
     
-    
-    contenedorCarritoVacio.classList.add("disabled");
-    contenedorCarritoProductos.classList.add("disabled");
-    contenedorCarritoAcciones.classList.add("disabled");
-    contenedorCarritoComprado.classList.remove("disabled");
-    
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+          
+          
+            contenedorCarritoVacio.classList.add("disabled");
+            contenedorCarritoProductos.classList.add("disabled");
+            contenedorCarritoAcciones.classList.add("disabled");
+            contenedorCarritoComprado.classList.remove("disabled");
+
+        }
+
+        
+    })
 
 }
+
+
